% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_plot_data}
\alias{aggregate_plot_data}
\title{Aggregate DGP data to prepare for simple time series plot}
\usage{
aggregate_plot_data(data)
}
\arguments{
\item{data}{The output dataset of DGP function}
}
\value{
Aggregated data ready for ggplot use
}
\description{
Aggregate DGP data to prepare for simple time series plot
}
\keyword{Internal}
