% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{att_group}
\alias{att_group}
\title{Calculates the average treatment effect of the treated (ATT)}
\usage{
att_group(
  data,
  treatment_effect = "treatment_effect",
  group = "group",
  time = "time"
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{treatment_effect}{The name of the column in \code{data} that contains the true individual treatment effects (default \code{"treatment_effect"}).}

\item{group}{The name of the column in \code{data} that contains the group indicators (default \code{"group"}).}

\item{time}{The name of the column in \code{data} that contains the time indicators (default \code{"time"}).}
}
\value{
A numeric value representing the ATT_group.
}
\description{
Calculates the average treatment effect of the treated (ATT)
}
