% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{DGP_full}
\alias{DGP_full}
\title{A flexible DGP function to generate panel data with differing time trends and treatment
effects between groups.}
\usage{
DGP_full(
  number_observations,
  time_periods,
  group_number,
  group_te,
  time_of_treatment,
  trend_vectors,
  group_level = NULL,
  exposure_list = NULL,
  group_share = "even",
  s_d = 0,
  error_spread = 0
)
}
\arguments{
\item{number_observations}{The number of observations in the dataset.}

\item{time_periods}{The number of time periods over which the observations are
observed.}

\item{group_number}{The number of groups in the sample. A group corresponds to
a distinct combination of treatment time and time trend.}

\item{group_te}{A vector indicating the treatment effect size for each group.
The length should equal the number of groups.}

\item{time_of_treatment}{A vector of length group_number, indicating in which period
each group first receives treatment. Never treeated groups are marked with -1.}

\item{trend_vectors}{A list of vectors indicating the time trend for each group.
Each vector should be of length time_periods, and the list of length group_numbers.
Alternatively a single vector can be passed - the function will use this time trend for all groups.
Vectors of common time trends can conveniently be created with the inbuilt trend functions.}

\item{group_level}{The level of Y at time 0. Note that the time trends can also
introduce a Y level above 0, but the trend fucntions in this package do not.}

\item{exposure_list}{A vector indicating the development of the treatment effect
after the initial treatment as a fraction of the respective group treatment effect
or a list of vectors indicating the development for each group individually.
Each vector should be of length time_periods - min(time_of_treatment).
A vector c(1,0.5, 0, 0, (...)) would indicate a treatmente effect of 1* group_te in the
treatment period, 0.5*group_te in the first period after treatment, and 0 after.}

\item{group_share}{A list of length group_number, which indicates how many obsevations
are in each group. The list entries should add up to the total number of observations.
The default "even" will create equally sized groups.}

\item{s_d}{The standard deviation of the normal distribution each individual observation
is drawn from. Note this is independent of group level noise, which may be introduced
via the trend functions.}
}
\description{
A flexible DGP function to generate panel data with differing time trends and treatment
effects between groups.
}
