% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{simple_ts_plot}
\alias{simple_ts_plot}
\title{Create simple time series plot:}
\usage{
simple_ts_plot(agg_data, demean)
}
\arguments{
\item{agg_data}{Aggregated plot Data}
}
\value{
A ggplot object
}
\description{
Create simple time series plot:
}
\keyword{Internal}
