% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthpanel-package.R
\docType{package}
\name{synthpanel-package}
\alias{synthpanel}
\alias{synthpanel-package}
\title{synthpanel: Simulating panel data}
\description{
Convenience functions to ease the simulation of panel data sets and common problems they pose in a causal inference setting.
}
\author{
\strong{Maintainer}: Fewo Nett \email{nettfewo@gmail.com}

}
\keyword{internal}
