% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{treatment_dgp_hetero}
\alias{treatment_dgp_hetero}
\title{Create a treatment vector, depending on time, group, and exposure.}
\usage{
treatment_dgp_hetero(
  number_units,
  time_periods,
  group_number,
  group_share,
  time_of_treatment,
  group_te,
  exposure_list
)
}
\description{
Create a treatment vector, depending on time, group, and exposure.
}
\keyword{Internal}
