% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{trend_brownian_motion}
\alias{trend_brownian_motion}
\title{Create a time trend based on drifting brownian motion.}
\usage{
trend_brownian_motion(time_periods, drift = 0, s_d = 1)
}
\arguments{
\item{time_periods}{The number of time periods over which the time trend should
be generated.}

\item{s_d}{The standard deviation of the normal distribution. The higher it is, the
more volatile the trend.}
}
\value{
A numerical vector of length time_periods
}
\description{
Create a time trend based on drifting brownian motion.
}
\examples{
time_periods <- 5
noise_sd <- 0.5
trend_brownian_motion(time_periods)

}
