% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{trend_linear}
\alias{trend_linear}
\title{Create a simple linear time trend}
\usage{
trend_linear(time_periods, slope)
}
\arguments{
\item{time_periods}{The number of time periods over which the time trend should
be generated.}

\item{slope}{The slope of the linear function}
}
\value{
A numerical vector
}
\description{
Create a simple linear time trend
}
\examples{
time_periods <- 5
slope <- 0.5
trend_linear(time_periods, slope)

}
