% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{trend_linear_noisy}
\alias{trend_linear_noisy}
\title{Create a linear time trend with normally distributed error nois in each period.}
\usage{
trend_linear_noisy(time_periods, slope, noise_mean, noise_sd)
}
\arguments{
\item{time_periods}{The number of time periods over which the time trend should
be generated.}

\item{slope}{The slope of the linear function}

\item{noise_mean}{The mean of the normal distribution. Should be 0 in most cases.}

\item{noise_sd}{The standard deviation of the normal distribution. The higher it is, the
more volatile the trend}
}
\value{
A numerical vector of length time_periods
}
\description{
Create a linear time trend with normally distributed error nois in each period.
}
\examples{
time_periods <- 5
slope <- 0.5
noise_mean <- 0
noise_sd <- 0.5
trend_linear_noisy(time_periods, slope, noise_mean, noise_sd)

}
