% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{unit_dgp_base}
\alias{unit_dgp_base}
\title{Creates a small dataframe that allocates units into groups and adds the time trends}
\usage{
unit_dgp_base(
  number_observations,
  time_periods,
  trend_vectors,
  group_vector_long,
  group_level,
  group_share
)
}
\description{
Creates a small dataframe that allocates units into groups and adds the time trends
}
\keyword{Internal}
