% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_dyn}
\alias{weights_dyn}
\title{Calculate dynamic weights}
\usage{
weights_dyn(data, group, time, treat_indicator)
}
\arguments{
\item{data}{The dataset.}

\item{group}{The name of the group variable column}

\item{time}{The name of the time variable column}

\item{treat_indicator}{The name of the treatment indicator column}
}
\value{
A numerical vector that contains the weights for each observation in the passed dataset
}
\description{
Calculate dynamic weights
}
