% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{att_group}
\alias{att_group}
\title{Calculates the average treatment effect of the treated (ATT)}
\usage{
att_group(treatment_effect, group, time)
}
\arguments{
\item{treatment_effect}{The vector of true individual treatment effects as generated
by the DGP functions.}

\item{group}{The group indicator vector as produced by the DGP functions.}

\item{time}{The time indicator vector as produced by the DGP functions.}
}
\value{
An integer representing the ATT.
}
\description{
Calculates the average treatment effect of the treated (ATT)
}
