% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{att_simple}
\alias{att_simple}
\title{Calculates the simple average treatment effect of the treated (ATT_simple)}
\usage{
att_simple(data, treatment_effect = "treatment_effect", treat = "treat")
}
\arguments{
\item{data}{A data frame containing the data.}

\item{treatment_effect}{The name of the column in \code{data} that contains the true individual treatment effects (default \code{"treatment_effect"}).}

\item{treat}{The name of the column in \code{data} that indicates whether the unit is treated (1) or not (0) in a given time period (default \code{"treat"}).}
}
\value{
A numeric value representing the ATT_simple.
}
\description{
Calculates the simple average treatment effect of the treated (ATT_simple)
}
