% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{get_all_ests}
\alias{get_all_ests}
\title{Compute all estimates}
\usage{
get_all_ests(
  data,
  ests = "all",
  Y = "Y",
  treat = "treat",
  time = "time",
  unit = "unit",
  group = "group",
  control = "notyettreated",
  aggregation = "simple",
  cluster_by = NA,
  iteration = NA
)
}
\description{
Compute all estimates
}
