% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{group_te}
\alias{group_te}
\title{Calculates the true treatment effect on a group level}
\usage{
group_te(
  data,
  treatment_effect = "treatment_effect",
  group = "group",
  time = "time"
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{treatment_effect}{The name of the column in \code{data} that contains the true individual treatment effects (default \code{"treatment_effect"}).}

\item{group}{The name of the column in \code{data} that contains the group indicators (default \code{"group"}).}

\item{time}{The name of the column in \code{data} that contains the time indicators (default \code{"time"}).}
}
\value{
A data frame that presents mean treatment effect by group.
}
\description{
Calculates the true treatment effect on a group level
}
