% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{group_time_te}
\alias{group_time_te}
\title{Calculates the true treatment effect on a group x time level}
\usage{
group_time_te(treatment_effect, group, time)
}
\arguments{
\item{treatment_effect}{The vector of true individual treatment effects as generated
by the DGP functions.}

\item{group}{The group indicator vector as produced by the DGP functions.}

\item{time}{The time indicator vector as produced by the DGP functions.}
}
\value{
A data frame that presents mean treatment effect by time x group.
}
\description{
Calculates the true treatment effect on a group x time level
}
