% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Create a simple time series plot of the average Y over time by group.}
\usage{
time_series_plot(Y, time, group, demean = FALSE, return_data = FALSE)
}
\arguments{
\item{Y}{The outcome variable.}

\item{time}{Time period of the respective data point.}

\item{group}{The treatment group of the respective data point.}

\item{demean}{If true, Y will be demeaned in period 1.}

\item{return_data}{If true, the function will return the aggregated data instead of the plot object.}
}
\value{
Returns a ggplot object, or the aggregated data if return_data is set to TRUE.
}
\description{
Create a simple time series plot of the average Y over time by group.
}
